added forbid screenshots logic
