--- android/src/main/java/io/parity/signer/components/ComposeUtils.kt
@@ -1,17 +1,0 @@
-package io.parity.signer.components
-
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.DisposableEffect
-import androidx.compose.ui.platform.LocalView
-
-
-@Composable
-fun KeepScreenOn() {
-	val currentView = LocalView.current
-	DisposableEffect(Unit) {
-		currentView.keepScreenOn = true
-		onDispose {
-			currentView.keepScreenOn = false
-		}
-	}
-}
