+++ android/src/main/java/io/parity/signer/models/UiUtils.kt
@@ -1,5 +1,8 @@
 package io.parity.signer.models
 
+import android.app.Activity
+import android.content.Context
+import android.content.ContextWrapper
 import androidx.compose.ui.Modifier
 
 
@@ -10,3 +13,18 @@ fun Modifier.conditional(condition : Boolean, modifier : Modifier.() -> Modifier
 		this
 	}
 }
+
+/**
+ * Method, used in accompanist/permission by Google
+ */
+fun Context.findActivity(): Activity? {
+	var context = this
+	while (context is ContextWrapper) {
+		if (context is Activity) return context
+		context = context.baseContext
+	}
+	return null
+}
+
+
+
